+++ test/unit_tests.cpp
@@ -1,6 +1,6 @@
 #define CATCH_CONFIG_MAIN
 #include "catch.hpp"
+#include "data_reader_tests.h"
 #include "exact_ops.h"
-#include "input_manager_tests.h"
 #include "map_matrix_tests.h"
 #include "serialization_tests.h"
